@@ -1,10 +1,10 @@
 package com.diconium.oakgit.engine.model;
 
 import com.diconium.oakgit.jdbc.OakGitResultSet;
+import com.diconium.oakgit.jdbc.util.SqlType;
 import lombok.Getter;
 import lombok.NonNull;
 import lombok.Setter;
-import org.apache.calcite.avatica.SqlType;
 import org.apache.commons.lang3.StringUtils;
 
 import java.util.function.Consumer;
