@@ -1,6 +1,7 @@
 package com.diconium.oakgit.engine.model;
 
 import com.diconium.oakgit.jdbc.OakGitResultSet;
+import com.diconium.oakgit.jdbc.util.SqlType;
 import com.diconium.oakgit.queryparsing.QueryAnalyzer;
 import com.diconium.oakgit.queryparsing.SingleValueId;
 import lombok.Getter;
@@ -8,7 +9,6 @@ import lombok.NonNull;
 import lombok.Setter;
 import lombok.ToString;
 import net.sf.jsqlparser.statement.Statement;
-import org.apache.calcite.avatica.SqlType;
 import org.apache.commons.lang3.StringUtils;
 
 import java.util.Map;
@@ -44,6 +44,45 @@ public class NodeAndSettingsEntry implements ContainerEntry<NodeAndSettingsEntry
 
     private byte[] bdata;
 
+    public static NodeAndSettingsEntry buildNodeSettingsDataForInsert(@NonNull Statement statement, @NonNull Map<Integer, Object> placeholderData, @NonNull QueryAnalyzer analyzer) {
+        NodeAndSettingsEntry data = new NodeAndSettingsEntry()
+            .setId(analyzer.getId(statement, placeholderData).orElse(SingleValueId.INVALID_ID).value());
+        analyzer
+            .getDataField(statement, "MODIFIED", Long.class, placeholderData)
+            .ifPresent(data::setModified);
+        analyzer
+            .getDataField(statement, "HASBINARY", Integer.class, placeholderData)
+            .ifPresent(data::setHasBinary);
+        analyzer
+            .getDataField(statement, "DELETEDONCE", Integer.class, placeholderData)
+            .ifPresent(data::setDeletedOnce);
+        analyzer
+            .getDataField(statement, "MODCOUNT", Long.class, placeholderData)
+            .ifPresent(data::setModCount);
+        analyzer
+            .getDataField(statement, "CMODCOUNT", Long.class, placeholderData)
+            .ifPresent(data::setCModCount);
+        analyzer
+            .getDataField(statement, "DSIZE", Long.class, placeholderData)
+            .ifPresent(data::setDSize);
+        analyzer
+            .getDataField(statement, "VERSION", Integer.class, placeholderData)
+            .ifPresent(data::setVersion);
+        analyzer
+            .getDataField(statement, "SDTYPE", Integer.class, placeholderData)
+            .ifPresent(data::setSdType);
+        analyzer
+            .getDataField(statement, "SDMAXREVTIME", Long.class, placeholderData)
+            .ifPresent(data::setSdMaxRevTime);
+        analyzer
+            .getDataField(statement, "DATA", String.class, placeholderData)
+            .ifPresent(f -> data.setData(f.getBytes()));
+        analyzer
+            .getDataField(statement, "BDATA", String.class, placeholderData)
+            .ifPresent(f -> data.setBdata(f.getBytes()));
+        return data;
+    }
+
     public NodeAndSettingsEntry appendData(byte[] addedData) {
         byte[] appendedData = new byte[data.length + addedData.length];
         System.arraycopy(data, 0, appendedData, 0, data.length);
@@ -51,45 +90,6 @@ public class NodeAndSettingsEntry implements ContainerEntry<NodeAndSettingsEntry
 
         this.data = appendedData;
         return this;
-    }
-
-    public static NodeAndSettingsEntry buildNodeSettingsDataForInsert(@NonNull Statement statement, @NonNull Map<Integer, Object> placeholderData, @NonNull QueryAnalyzer analyzer) {
-        NodeAndSettingsEntry data = new NodeAndSettingsEntry()
-                .setId(analyzer.getId(statement, placeholderData).orElse(SingleValueId.INVALID_ID).value());
-        analyzer
-                .getDataField(statement, "MODIFIED", Long.class, placeholderData)
-                .ifPresent(data::setModified);
-        analyzer
-                .getDataField(statement, "HASBINARY", Integer.class, placeholderData)
-                .ifPresent(data::setHasBinary);
-        analyzer
-                .getDataField(statement, "DELETEDONCE", Integer.class, placeholderData)
-                .ifPresent(data::setDeletedOnce);
-        analyzer
-                .getDataField(statement, "MODCOUNT", Long.class, placeholderData)
-                .ifPresent(data::setModCount);
-        analyzer
-                .getDataField(statement, "CMODCOUNT", Long.class, placeholderData)
-                .ifPresent(data::setCModCount);
-        analyzer
-                .getDataField(statement, "DSIZE", Long.class, placeholderData)
-                .ifPresent(data::setDSize);
-        analyzer
-                .getDataField(statement, "VERSION", Integer.class, placeholderData)
-                .ifPresent(data::setVersion);
-        analyzer
-                .getDataField(statement, "SDTYPE", Integer.class, placeholderData)
-                .ifPresent(data::setSdType);
-        analyzer
-                .getDataField(statement, "SDMAXREVTIME", Long.class, placeholderData)
-                .ifPresent(data::setSdMaxRevTime);
-        analyzer
-                .getDataField(statement, "DATA", String.class, placeholderData)
-                .ifPresent(f -> data.setData(f.getBytes()));
-        analyzer
-                .getDataField(statement, "BDATA", String.class, placeholderData)
-                .ifPresent(f -> data.setBdata(f.getBytes()));
-        return data;
     }
 
     @Override
