@@ -29,8 +29,8 @@ public class SelectByIdAnalyzer implements QueryAnalyzer {
     @Override
     public boolean interestedIn(Statement statement) {
         return statement instanceof Select &&
-                ((Select) statement).getSelectBody() instanceof PlainSelect &&
-                ((PlainSelect) ((Select) statement).getSelectBody()).getWhere() instanceof EqualsTo;
+            ((Select) statement).getSelectBody() instanceof PlainSelect &&
+            ((PlainSelect) ((Select) statement).getSelectBody()).getWhere() instanceof EqualsTo;
     }
 
     @Override
@@ -60,16 +60,16 @@ public class SelectByIdAnalyzer implements QueryAnalyzer {
 
         if (id.isPresent()) {
             return new SelectFromContainerByIdCommand()
-                    .setOriginSql(statement.toString())
-                    .setPlaceholderData(placeholderData)
-                    .setContainerName(getTableName(statement))
-                    .setId(id.get().value());
+                .setOriginSql(statement.toString())
+                .setPlaceholderData(placeholderData)
+                .setContainerName(getTableName(statement))
+                .setId(id.get().value());
         }
 
         return new ErrorCommand()
-                .setOriginSql(statement.toString())
-                .setPlaceholderData(placeholderData)
-                .setErrorMessage("Cannot create command with an invalid id");
+            .setOriginSql(statement.toString())
+            .setPlaceholderData(placeholderData)
+            .setErrorMessage("Cannot create command with an invalid id");
     }
 
     private Optional<QueryId> extractIdFromWhere(Expression whereExpression, Map<Integer, Object> placeholderData) {
